backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
crewai>=0.80.0
crewai-tools>=0.17.0
duckduckgo-search>=6.0.0
llama-index-core>=0.11.0
llama-index-readers-web>=0.2.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.28.0
